Added eeprom and some upload functions

+++ DMX_Merger/Middleware/EEPROM_24xxxx/eeprom.c
@@ -16,10 +16,10 @@
 /******************************************************************************
 * Function Prototypes
 *******************************************************************************/
-static inline uint8_t writeValidation(eeprom_cfg_Config *config, uint16_t startAddr, uint8_t *data, uint8_t dataSize);
+static inline uint8_t writeValidation(eeprom_cfg_Config *config, uint16_t startAddr, uint8_t *data, uint16_t dataSize);
 static inline uint8_t readValidation(eeprom_cfg_Config *config, uint16_t startAddr, uint8_t *data, uint16_t dataSize);
 static inline void writeStartAddr(eeprom_cfg_Config *config, uint16_t startAddr);
-static inline void writeData(eeprom_cfg_Config *config, uint8_t *data, uint8_t dataSize);
+static inline void writeData(eeprom_cfg_Config *config, uint8_t *data, uint16_t dataSize);
 static inline void readData(eeprom_cfg_Config *config, uint8_t *data, uint16_t dataSize);
 static inline void callbackError(uint8_t type);
 /******************************************************************************
@@ -34,7 +34,7 @@ void eeprom_init(eeprom_cfg_Config *config){
 	}
 }
 
-void eeprom_write(eeprom_cfg_Config *config, uint16_t startAddr, uint8_t *data, uint8_t dataSize){
+void eeprom_write(eeprom_cfg_Config *config, uint16_t startAddr, uint8_t *data, uint16_t dataSize){
 	if(writeValidation(config, startAddr, data, dataSize) == OK){
 		writeStartAddr(config, startAddr);
 		writeData(config, data, dataSize);
@@ -62,7 +62,7 @@ static inline void writeStartAddr(eeprom_cfg_Config *config, uint16_t startAddr)
 	config->memAddr = startAddr;
 }
 
-static inline void writeData(eeprom_cfg_Config *config, uint8_t *data, uint8_t dataSize){
+static inline void writeData(eeprom_cfg_Config *config, uint8_t *data, uint16_t dataSize){
 	config->dataSize = dataSize;
 	config->txBuffer = data;
 	config->callbackType = EEPROM_CFG_CALL_TX_START;
@@ -76,7 +76,7 @@ static inline void readData(eeprom_cfg_Config *config, uint8_t *data, uint16_t d
 	if(config->callback(config) != OK) callbackError(EEPROM_CFG_CALL_RX_ERROR);
 }
 
-static inline uint8_t writeValidation(eeprom_cfg_Config *config, uint16_t startAddr, uint8_t *data, uint8_t dataSize){
+static inline uint8_t writeValidation(eeprom_cfg_Config *config, uint16_t startAddr, uint8_t *data, uint16_t dataSize){
 	if(config->callback == NULL) return NOK;
 	if(data == NULL) return NOK;
 	if((startAddr + dataSize) > config->maxAddress) return NOK;
